refactor(dao): 함수형 인터페이스를 jdbcTemplate 내부로 이동
