--- src/main/java/database/dao/BatchPreparedStatementSetter.java
@@ -1,9 +1,0 @@
-package database.dao;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface BatchPreparedStatementSetter<T> {
-    void setValues(PreparedStatement ps, T item) throws SQLException;
-}
