--- src/main/java/database/dao/PreparedStatementCallback.java
@@ -1,11 +1,0 @@
-package database.dao;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface PreparedStatementCallback<T> {
-
-    T execute(PreparedStatement preparedStatement) throws SQLException;
-
-}
